Added deletion confirmation

+++ LeisureReviews/Scripts/profile.ts
@@ -1,18 +1,21 @@
 ﻿$('.delete-review-button').on('click', async function () {
     var currentButton = $(this)
-    showButtonSpinner(currentButton)
     let reviewId = $(this).parents('.review-card').data('id')
-    return await $.ajax(`/DeleteReview?reviewId=${reviewId}`, {
-        method: 'DELETE'
-    }).done(function (reviewId) {
-        $(document).find(`.review-card[data-id="${reviewId}"]`).fadeOut('slow')
-        //@ts-ignore
-        UIkit.notification({ message: 'The review was successfully deleted', status: 'success', pos: 'bottom-center' })
-    }).fail(function () {
-        //@ts-ignore
-        UIkit.notification({ message: 'An error occurred during review deleiton', status: 'danger', pos: 'bottom-center' })
-    }).always(function () {
-        hideButtonSpinner(currentButton)
+    //@ts-ignore
+    return await UIkit.modal.confirm('Are you sure you want to delete the review?').then(async function () {
+        showButtonSpinner(currentButton)
+        return await $.ajax(`/DeleteReview?reviewId=${reviewId}`, {
+            method: 'DELETE'
+        }).done(function (reviewId) {
+            $(document).find(`.review-card[data-id="${reviewId}"]`).fadeOut('slow')
+            //@ts-ignore
+            UIkit.notification({ message: 'The review was successfully deleted', status: 'success', pos: 'bottom-center' })
+        }).fail(function () {
+            //@ts-ignore
+            UIkit.notification({ message: 'An error occurred during review deleiton', status: 'danger', pos: 'bottom-center' })
+        }).always(function () {
+            hideButtonSpinner(currentButton)
+        })
     })
 })
